removed UiLength::realPixels; mac adaptations to switch to DIP
UiLength::realPixels has no clear use case that requires it to work as
a fundamental unit everywhere in the UI. When real pixels need to be
known (for example, for adapting image scaling etc.) then we can add a
specific API for finding out. There is no need to carry around the
baggage of having to be able to convert to and from realPixel
everywhere.
Also, the concept of “real pixel” does not even exist on all platforms.
For example, most web browsers only really work with DIPs and there is
no good way to find out what real pixels are.