axed the concept of "invisible" Dip differences
after careful consideration this concept was not really useful. To be
safe that differences are really "invisible" we have to choose a very
small limit. And it is unlikely that anything falls into that limit.
So instead we only deal with floating point inaccuracies rather than
trying to optimize away invisible overshoots.