ITextUi write functions are now synchronous (#52)
All write function of ITextUi are now synchronous. The additional complexity due to async I/O outweighs the small benefit of having this be non-blocking.
Also, this is intended to be used for actual user interaction, not for data piping between programs (in contrast to stdout, for example). If that is the case then the terminal program will continually read what we write, so no long blocking will occur.

Note that if no user is on the other side then blocking might occur. However, this can only happen for commandline apps that run in non-interactive mode. In that case it does not matter that we block because the app - in fact it is probably better.