documented overwrite behaviour of add()
BDN-35 #done documented overwrite behaviour of add(). Add remains a void
method - adding a return iterator would decrease performance
significantly when it is not needed (which is most of the time). If we
do need a return iterator in the future then we can add another method
with a different name,