Remove ui thread safety (#57)
Removed all code that had the task of making the UI components usable from any thread. It was decided that the small advantages multithread support in the UI are not worth the headaches.
The only remaining multithread support is the fact the View references can be released from any thread (and the destruction is automatically executed in the main thread).